refactor: extract date utility
