+++ pages/api/log.js
@@ -1,4 +1,5 @@
 const fs = require('fs');
+import { getCurrentDate } from '../../lib/dateUtils';
 
 export default async function handler(req, res) {
   // postじゃないならエラー
@@ -37,30 +38,3 @@ export default async function handler(req, res) {
   });
 }
 
-function getCurrentDate() {
-  const currentDate = new Date();
-  // 年
-  const year = currentDate.getFullYear();
-  // 月
-  const month = currentDate.getMonth() + 1;
-  // 日
-  const day = currentDate.getDate();
-  // 時間
-  const hour = currentDate.getHours();
-  // 分
-  const min = currentDate.getMinutes();
-  // 秒
-  const sec = currentDate.getSeconds();
-  // ミリ秒
-  const msec = currentDate.getMilliseconds();
-
-  const date =
-    year + '-' +
-    String(month).padStart(2, "0") + '-' +
-    String(day).padStart(2, "0") + '-' +
-    String(hour).padStart(2, "0") + '-' +
-    String(min).padStart(2, "0") + '-' +
-    String(sec).padStart(2, "0") + '-' +
-    String(msec).padStart(3, "0");
-  return date;
-}
